+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - ADD: Custom amount of formula to make
 - ADD: Solvent in ingredient level
 - ADD: Ingredients CSV import support
+- ADD: Formula quantity decimal step
 - FIX: Formula not showning properly when new
 - FIX: Modal window close on bg click